pyproject.toml
[project]
name = "py-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "devtools>=0.12.2",
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "langchain-text-splitters>=0.3.6",
    "logfire[asyncpg,fastapi,sqlite3]>=3.5.3",
    "pydantic-ai-slim[anthropic,groq,openai,vertexai]==0.0.21",
    "pymongo[srv]>=4.11.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "spacy>=3.8.4",
    "spacy-layout>=0.0.11",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
]
